+++ Sources/QAMenuUIKit/Internal/UI/Helper/SingleLineTextEditorController.swift
@@ -1,0 +1,127 @@
+//
+//  SingleLineTextEditorController.swift
+//
+//  Created by Hans Seiffert on 07.02.21.
+//
+//  ---
+//  MIT License
+//
+//  Copyright © 2021 Hans Seiffert
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in all
+//  copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// 
+
+import UIKit
+
+internal protocol SingleLineTextEditorControllerProtocol: TextEditorProtocol {
+
+    init(
+        title: String?,
+        text: String?,
+        onCancelEditing: @escaping (_ editor: SingleLineTextEditorControllerProtocol) -> Void,
+        onEndEditing: @escaping (_ newValue: String?, _ editor: SingleLineTextEditorControllerProtocol) -> Void
+    )
+}
+
+internal final class SingleLineTextEditorController: SingleLineTextEditorControllerProtocol {
+
+    // MARK: - Properties (Public)
+
+    internal var viewController: UIViewController {
+        return self.inputAlertController
+    }
+
+    // MARK: - Properties (Private)
+
+    private var title: String?
+    private var text: String?
+
+    private lazy var inputAlertController: UIAlertController = {
+        return self.makeAlertViewController()
+    }()
+
+    private let onCancelEditing: (_ editor: SingleLineTextEditorControllerProtocol) -> Void
+    private let onEndEditing: (_ newValue: String?, _ editor: SingleLineTextEditorControllerProtocol) -> Void
+
+    internal init(
+        title: String?,
+        text: String?,
+        onCancelEditing: @escaping (_ editor: SingleLineTextEditorControllerProtocol) -> Void,
+        onEndEditing: @escaping (_ newValue: String?, _ editor: SingleLineTextEditorControllerProtocol) -> Void
+    ) {
+        self.title = title
+        self.text = text
+        self.onCancelEditing = onCancelEditing
+        self.onEndEditing = onEndEditing
+    }
+
+    @available(*, unavailable)
+    internal required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    // MARK: - Methods
+
+    internal func dismiss() {
+        self.inputAlertController.dismiss(animated: true)
+    }
+
+    // MARK: - Helper (Private)
+
+    internal func makeAlertViewController() -> UIAlertController {
+        let alert = UIAlertController(
+            title: self.title,
+            message: "",
+            preferredStyle: .alert
+        )
+        alert.addTextField { [weak self] textField in
+            textField.text = self?.text
+            textField.addTarget(
+                self,
+                action: #selector(Self.textFieldDidChange(_:)),
+                for: .editingChanged
+            )
+        }
+
+        let cancelAction = UIAlertAction(
+            title: "Cancel",
+            style: .cancel,
+            handler: { [weak self] _ in
+                guard let self = self else { return }
+                self.onCancelEditing(self)
+            }
+        )
+        alert.addAction(cancelAction)
+
+        let saveAction = UIAlertAction(
+            title: "Save",
+            style: .default,
+            handler: { [weak self] _ in
+                guard let self = self else { return }
+                self.onEndEditing(self.text, self)
+            }
+        )
+        alert.addAction(saveAction)
+
+        return alert
+    }
+
+    @objc private func textFieldDidChange(_ textField: UITextField) {
+        self.text = textField.text
+    }
+}