+++ Tests/QAMenuUIKit-UnitTests/TestDoubles/SingleLineTextEditorControllerMock.swift
@@ -1,12 +1,12 @@
 //
-//  ItemUIRepresentableDelegateSpy.swift
+//  SingleLineTextEditorControllerMock.swift
 //
-//  Created by Hans Seiffert on 24.11.20.
+//  Created by Hans Seiffert on 13.02.21.
 //
 //  ---
 //  MIT License
 //
-//  Copyright © 2020 Hans Seiffert
+//  Copyright © 2021 Hans Seiffert
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -24,29 +24,33 @@
 //  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-//
+// 
 
-import XCTest
-import QAMenu
+import UIKit
 @testable import QAMenuUIKit
 
-class ItemUIRepresentableDelegateSpy: UIViewController, ItemUIRepresentableDelegate {
+internal final class AlertControllerMock: UIAlertController {}
 
-    var _present = [
-        (viewController: UIViewController, animated: Bool, completion: (() -> Void)?)
-    ]()
+internal final class SingleLineTextEditorControllerMock: SingleLineTextEditorControllerProtocol {
 
-    var _updateContainerHeightCount = 0
+    internal var _dismissCallCount = 0
 
-    var presentationContext: UIViewController {
-        return self
+    internal let onCancelEditing: (_ editor: SingleLineTextEditorControllerProtocol) -> Void
+    internal let onEndEditing: (_ newValue: String?, _ editor: SingleLineTextEditorControllerProtocol) -> Void
+
+    internal var viewController: UIViewController = AlertControllerMock()
+
+    internal init(
+        title: String?,
+        text: String?,
+        onCancelEditing: @escaping (SingleLineTextEditorControllerProtocol) -> Void,
+        onEndEditing: @escaping (String?, SingleLineTextEditorControllerProtocol) -> Void
+    ) {
+        self.onCancelEditing = onCancelEditing
+        self.onEndEditing = onEndEditing
     }
 
-    func updateContainerHeight(for elementView: ItemUIRepresentable?) {
-        self._updateContainerHeightCount += 1
-    }
-
-    override func present(_ viewControllerToPresent: UIViewController, animated flag: Bool, completion: (() -> Void)? = nil) {
-        self._present.append((viewControllerToPresent, flag, completion))
+    internal func dismiss() {
+        _dismissCallCount += 1
     }
 }